refactor(GameBoard,Ladder) : 포인트 순회 로직 GameBoard -> Ladder 로 이동
- 이에 대한 테스트 코드 작성

-[Command 관련 피드백 링크](https://github.com/woowacourse/java-ladder/pull/365#discussion_r1507567994)
